refactor: dropdown 외부 클릭 로직
