+++ src/pages/selectSchdule/SelectPage.tsx
@@ -48,60 +48,9 @@ function SelectPage() {
     setScheduleList(scheduleList.filter((item) => item.id !== index));
   };
 
-  const handleDropDown = (index: number) => {
-    setDateList((prevDateList) => {
-      return prevDateList.map((item) => {
-        if (item.id === index) {
-          return {
-            ...item,
-            dropDown: !item.dropDown,
-          };
-
-          return item;
-        }
-      });
-    });
-  };
-
-  // const handleStartDropDown = (index: number) => {
-  //   setDateList((prevDateList) => {
-  //     return prevDateList.map((item) => {
-  //       if (item.id === index) {
-  //         return {
-  //           ...item,
-  //           dropDown: true, // Set the desired value for dropDown
-  //           startTimeDropDown: !item.startTimeDropDown,
-  //           endTimeDropDown: true,
-  //         };
-  //       }
-  //       return item;
-  //     });
-  //   });
-  // };
-
-  // const handleEndDropDown = (index: number) => {
-  //   setDateList((prevDateList) => {
-  //     return prevDateList.map((item) => {
-  //       if (item.id === index) {
-  //         return {
-  //           ...item,
-  //           dropDown: true, // Set the desired value for dropDown
-  //           startTimeDropDown: true,
-  //           endTimeDropDown: !item.endTimeDropDown,
-  //         };
-  //       }
-  //       return item;
-  //     });
-  //   });
-  // };
-
   return (
     <SelectPageWrapper>
-      <SelectSchedule
-        dataList={dateList}
-        deleteData={deleteDataList}
-        handleDropDown={handleDropDown}
-      />
+      <SelectSchedule dataList={dateList} deleteData={deleteDataList} />
       <PlusButton onClick={addDateList} type="button">
         <PlusIc />
       </PlusButton>
