+++ src/pages/selectSchdule/components/DateDropDown.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useRef, SetStateAction, Dispatch } from 'react';
+import React from 'react';
 
 import Text from 'components/atomComponents/Text';
 import { useRecoilState } from 'recoil';
@@ -13,35 +13,19 @@ interface PropTypes {
   month: number;
   day: number;
   dayOfWeek: string;
-  handleDropDown: (id: number) => {};
-  setIsOpen: Dispatch<SetStateAction<boolean>>;
-  isOpen: boolean;
 }
 
-function DateDropDown({ month, day, dayOfWeek, setIsOpen, isOpen }: PropTypes) {
+function DateDropDown({ month, day, dayOfWeek }: PropTypes) {
   const [schedule, setSchedule] = useRecoilState(scheduleAtom);
-  const ref = useRef<HTMLDivElement>(null);
+
   const getDate = () => {
     const updatedDate = `${month}월 ${day}일 ${dayOfWeek}요일`;
     setSchedule({ ...schedule, date: updatedDate });
     console.log(updatedDate);
   };
-  useEffect(
-    () => {
-      const clickOutSide = (e: MouseEvent) => {
-        if (isOpen && ref.current && !ref.current.contains(e.target as Node)) {
-          setIsOpen(false);
-        }
-      };
-      document.addEventListener('mousedown', clickOutSide);
-      return () => {
-        document.removeEventListener('mousedown', clickOutSide);
-      };
-    },
-    [isOpen, ref.current],
-  );
+
   return (
-    <DropDownList onClick={getDate} ref={ref}>
+    <DropDownList onClick={getDate}>
       <Text font="button1" color={`${theme.colors.white}`}>
         {`${month}월 ${day}일 ${dayOfWeek}요일`}
       </Text>
